requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
GoDaddyPy==2.3.4
idna==3.4
pif==0.8.2
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
